+++ mesonbuild/compilers/c.py
@@ -40,10 +40,11 @@ from .compilers import (
 
 if T.TYPE_CHECKING:
     from ..coredata import KeyedOptionDictType
-    from ..dependencies import Dependency, ExternalProgram
+    from ..dependencies import Dependency
     from ..envconfig import MachineInfo
     from ..environment import Environment
     from ..linkers import DynamicLinker
+    from ..programs import ExternalProgram
 
     CompilerMixinBase = Compiler
 else:
