+++ mesonbuild/compilers/cpp.py
@@ -42,10 +42,11 @@ from .mixins.emscripten import EmscriptenMixin
 
 if T.TYPE_CHECKING:
     from ..coredata import KeyedOptionDictType
-    from ..dependencies import Dependency, ExternalProgram
+    from ..dependencies import Dependency
     from ..envconfig import MachineInfo
     from ..environment import Environment
     from ..linkers import DynamicLinker
+    from ..programs import ExternalProgram
     from .mixins.clike import CLikeCompiler as CompilerMixinBase
 else:
     CompilerMixinBase = object
